db fixes

+++ public_html/project/order_process.php
@@ -81,13 +81,8 @@ $stmt = $db->prepare("INSERT INTO Orders (user_id, total, money_recieved,payment
 $stmt->execute([":UID" => $user_id, ":total" => $total, ":money" => $total, ":payment_method" => $payment_type, ":place" => $Address]);
 $user_id = get_user_id();
 
-
-
-if (!$hasError) {
-
-    $db->commit();
-    $stmt = $db->prepare("SELECT max(id) as order_id FROM Orders");
-    $next_order_id = 0;
+$stmt = $db->prepare("SELECT max(id) as order_id FROM Orders");
+$next_order_id = 0;
     //get next order id
     try {
         $stmt->execute();
@@ -99,19 +94,6 @@ if (!$hasError) {
         $db->rollback();
     }
     
-    
-    if ($next_order_id > 0) {
-        $stmt = $db->prepare("UPDATE Products 
-        set stock = stock - (select IFNULL(quantity, 0) FROM JG_Cart WHERE item_id = Products.id and user_id = :uid) 
-        WHERE id in (SELECT item_id from 'JG_Cart where user_id = :uid)");
-        try {
-            $stmt->execute([":uid" => $user_id]);
-        } catch (PDOException $e) {
-            error_log("Update stock error: " . var_export($e, true));
-       
-            $next_order_id = 0; //using as a controller
-        }
-    }
     if ($next_order_id > 0) {
         $stmt = $db->prepare("INSERT INTO Orderitems (item_id, user_id, quantity, cost, order_id) 
         SELECT item_id, user_id, JG_Cart.quantity, cost, :order_id FROM JG_Cart JOIN Products on JG_Cart.item_id = Products.id
@@ -124,6 +106,25 @@ if (!$hasError) {
             $next_order_id = 0; //using as a controller
         }
     } 
+if (!$hasError) {
+
+    $db->commit();
+
+
+    
+    /*if ($next_order_id > 0) {
+        $stmt = $db->prepare("UPDATE Products 
+        set stock = stock - (select IFNULL(quantity, 0) FROM JG_Cart WHERE item_id = Products.id and user_id = :uid) 
+        WHERE id in (SELECT item_id from 'JG_Cart where user_id = :uid)");
+        try {
+            $stmt->execute([":uid" => $user_id]);
+        } catch (PDOException $e) {
+            error_log("Update stock error: " . var_export($e, true));
+       
+            $next_order_id = 0; //using as a controller
+        }
+    }*/
+ 
     $stmt = $db->prepare("DELETE FROM JG_Cart where user_id =  :userid");
 $stmt->execute([":userid" => $user_id]);
